debug: debugging action to see if im doing this right
