+++ Scripts/.gitignore
@@ -1,2 +1,4 @@
 output
 input
+
+__pycache__
